CadSoft Eagle 6 Schematic source and PDF for the Intel 4004 CPU
This is a single-page schematic containing the entire Intel 4004 CPU.
It was derived from the schematic bitmap provided in Lagos Kintli's
simulator.

This schematic necessarily carries the same Creative Commons non-
commercial license that applies to the rest of these materials.
See the files doc/License.txt and doc/Intel_IPNC_License.pdf for
additional details.